+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,0 +1,20 @@
+package lotto;
+
+import lotto.dto.LottoNumber;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+public class LottoNumberTest {
+
+    @ParameterizedTest
+    @ValueSource(ints = {0, 46})
+    @DisplayName("로또 숫자가 1 ~ 45에 해당하지 않으면 에러가 발생한다.")
+    void lottoNumberRange(int number) {
+        assertThatThrownBy(() -> {
+            new LottoNumber(number);
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+}
